backend2/app/requirements.txt
fastapi
uvicorn
google-generativeai
scikit-learn
numpy
pandas
pyttsx3
python-dotenv
pydantic
